chore: [#188448903] move modifyCurrentBusiness fn out of test folder
