+++ api/src/domain/tax-filings/taxFilingsInterfaceFactory.ts
@@ -1,7 +1,7 @@
 import { TaxFilingClient, TaxFilingInterface } from "@domain/types";
 import { fetchMunicipalityByName } from "@domain/user/fetchMunicipalityByName";
 import { getCurrentBusiness } from "@shared/domain-logic/getCurrentBusiness";
-import { modifyCurrentBusiness } from "@shared/test";
+import { modifyCurrentBusiness } from "@shared/domain-logic/modifyCurrentBusiness";
 import { UserData } from "@shared/userData";
 import dayjs from "dayjs";
